flask_backend/app/requirements.txt
Flask==2.1.2
jsonpickle==2.1.0
matplotlib==3.5.2
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.1
scikit_learn==1.1.1
seaborn==0.11.2
pickl==0.3.0
